+++ src/api/endpoints/tokens/get-tokens-details/v2.ts
@@ -63,10 +63,7 @@ export const getTokensDetailsV2Options: RouteOptions = {
         ),
       sortBy: Joi.string().valid("floorAskPrice", "topBidValue").default("floorAskPrice"),
       limit: Joi.number().integer().min(1).max(50).default(20),
-      continuation: Joi.alternatives().try(
-        Joi.string().pattern(/^((\d+|null)_\d+|\d+)$/),
-        Joi.string().pattern(base64Regex)
-      ),
+      continuation: Joi.string().pattern(base64Regex),
     })
       .or("collection", "contract", "token", "tokenSetId")
       .oxor("collection", "contract", "token", "tokenSetId")
@@ -370,14 +367,11 @@ export const getTokensDetailsV2Options: RouteOptions = {
         continuation = buildContinuation(continuation);
       }
 
-      const result = rawResult.map(async (r) => {
-        const sources = new Sources();
+      const sources = await Sources.getInstance();
+
+      const result = rawResult.map((r) => {
         const source = r.floor_sell_source_id
-          ? await sources.get(
-              fromBuffer(r.floor_sell_source_id),
-              fromBuffer(r.contract),
-              r.token_id
-            )
+          ? sources.get(fromBuffer(r.floor_sell_source_id), fromBuffer(r.contract), r.token_id)
           : null;
 
         return {