+++ src/api/endpoints/tokens/get-tokens-floor/v1.ts
@@ -10,6 +10,10 @@ import { formatEth, toBuffer } from "@/common/utils";
 const version = "v1";
 
 export const getTokensFloorV1Options: RouteOptions = {
+  cache: {
+    privacy: "public",
+    expiresIn: 5000,
+  },
   description: "Get the current best price of every on sale token in a collection",
   notes:
     "This API will return the best price of every token in a collection that is currently on sale",