+++ src/jobs/index.ts
@@ -11,6 +11,7 @@ import "@/jobs/order-updates";
 import "@/jobs/orderbook";
 import "@/jobs/token-updates";
 import "@/jobs/daily-volumes";
+import "@/jobs/update-attribute";
 
 // Export all job queues for monitoring through the UI.
 
@@ -31,6 +32,9 @@ import * as orderbookOrders from "@/jobs/orderbook/orders-queue";
 import * as orderbookTokenSets from "@/jobs/orderbook/token-sets-queue";
 import * as tokenUpdatesMintQueue from "@/jobs/token-updates/mint-queue";
 import * as dailyVolumes from "@/jobs/daily-volumes/daily-volumes";
+import * as handleNewSellOrder from "@/jobs/update-attribute/handle-new-sell-order";
+import * as handleNewBuyOrder from "@/jobs/update-attribute/handle-new-buy-order";
+import * as resyncAttributeCache from "@/jobs/update-attribute/resync-attribute-cache";
 
 export const allJobQueues = [
   arweaveSyncBackfill.queue,
@@ -50,4 +54,7 @@ export const allJobQueues = [
   orderbookTokenSets.queue,
   tokenUpdatesMintQueue.queue,
   dailyVolumes.queue,
+  handleNewSellOrder.queue,
+  handleNewBuyOrder.queue,
+  resyncAttributeCache.queue,
 ];